First commit
One client and multiple nodes. 
Assumption:  node is always the leader and client send message to that node until there is no message in the client size 
What needs to work on now; Leader election + data of the node?  (I could not find the nested structure for now, so we have to list everything out)