Update raft.mcrl2
Rename  ll to last log index
New to me: 
If 
if like this I need to use else inorder to work
define rhead for empty log to be yellow. So the log is empty then the term is 0 and the log index  is also 0. We assume that index start from 1 and also the term

TODO still need to run some tests, but now everything seems to run 
I have to go to play soccer now